requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
openai==0.27.0
python-dotenv==1.0.0
jinja2==3.1.2
